refactor: update translator to support multiple variables in single environment
change single static environment record to a list of variable names, apply-senv returns a
a pair of (depth . position) as a variable location, depth refers to index of target
environment record in environment, position refers to index of variable in environment record.

update nameless-var-exp to container (depth . position) pair.

exer 3.41